Monkeypatch TagBase.slugify to allow non ascii chars (cf #118)
It's possible to override the Tag model used by taggit, but this
means customizing two models, and so having new table in database
(or hardcoding their name to match the current taggit ones). And I
prefer not to go such a complex way for only overriding a slugify
method.
Let's wait for https://github.com/alex/django-taggit/pull/273 or
any other option without impact on DB.