Full refactor of User management
WARNING: this reset the migrations!
I've made this choice because at the moment we have only one
box with this server in production, which is the Azraq Camp box,
and I'm in Azraq right now so I'm able to manage this tricky
migration by hand, instead of adding it in the history.

The main idea of this change is that we don't have anymore one user
per box, but a generic user with all needed fields. Instead, every box
will define which field it want to use among the available ones.=