Do not fail if key is not in SelectMultipleField choices
This is a bug report from Azraq camp. Not sure this can happen.
It can happen in theory if the choices change, but this should
not be the case there. So I need to investigate as soon as I
have an SSH access to their server. For now, I hope this will
fix the 500 on the user detail view.