tests: Don't run coverage by default
The coverage report takes a lot of space, which can distract from the
actual test results.

This commit makes it so that running "py.test" or "make test" doesn't
check for coverage any more at all.

A new "make testcov" target is introduced, which both runs the tests and
reports coverage.