wifi: Get the Wi-Fi device
This assumes that the Wi-Fi was already enabled.

For now, we're only going to support the case where we have a single
NM-managed wireless device.

No wireless device is the obvious case where we can't do anything.

Multiple NM-managed wireless devices are more tricky, and it's a case
which is not going to happen any time soon, so we can figure it out when
that happens.