services: Wrap systemd units
We are going to revamp our system service management, basing it on
systemd.

Our target for deployment is Debian Jessie, so we can rely on systemd
being present on the host.

In systemd, services are a special case of the more generic "unit"
concept, so if we want to be able to manage services, we might as well
handle units.

This commit adds a new class, which wraps the systemd DBus API so that
we can handle units as objects in Python.