services: Introduce the systemd manager
This wraps the Manager object exposed by systemd on DBus.

This allows us to get a service by its user-facing name (e.g "nginx")
get the unit id as far as systemd is concerned (e.g "nginx.service"),
and finally get an instance of the Unit class introduced in the
previous commit.