services: Allow controlling a service state
We want to keep things simple for Ideascube: either a service is on or
it is off.

That means we couple together things that systemd considers separate:

* starting and enabling a service
* stopping and disabling a service

When an administrator stops a service to save resources, they don't want
it to be automatically started at next boot.

This commit also adds the possibility to directly restart a service,
because merely going through stop then start would also go through
disable then re-enable.

systemd offers a nice DBus API to control unit states, but unfortunately
we can't use it.

That is because that API was only made open through PolicyKit in systemd
216 and later, but Debian Jessie (our target for deployments) only has
systemd 215.

As a result, we have to settle down for running the systemctl command as
a subprocess.

However, we pass it through pkexec rather than sudo, which offers a
couple of advantages:

* without any configuration, PolicyKit will ask for the local
  administrator platform through the system authorization agent (for
  example a nice popup in GNOME), which makes things simple for
  developer workstations;
* configuring PolicyKit to grant the permissions without password (for
  deployments) is done by dropping a simple configuration file in a
  folder, which easier than editing the sudoers file.