services: Port the view to systemd
Now that we have a nice systemd wrapper to manage services, we can use
it in the view.

This change has a few benefits.

First, we can drop the custom "status_caller" thing, because systemd
knows how to tell if a service is active or not.

Second, we can completely drop the ideascube.serveradmin.utils module,
since all it did is now handled by the systemd wrapper.

Third, since the systemd module uses PolicyKit, then the web app
benefits from this as well. (see a previous commit message on what this
brings us)

Finally, the view will now not only stop running services, but also
disable them so that they are not started at next boot. The same goes
for starting inactive services and enabling them.

The view now better handles errors as well, reporting them to the admin
user.

The view was also entirely untested, which is now fixed with this
commit.

A few things to note: this port was made with minimal changes, in the
goal to quickly move to systemd and get the enable/disable feature which
was lacking.

In particular, this commit does not implement the new UI as mocked up.
But this can come in a future change.