tests: Reuse the same FakePopen
In a previous commit, I have added a FakePopen class for the systemd
unit tests.

That class is more capable than the FakePopen we already had in there,
so let's just reuse it and drop the old one, as it does everything
needed for the tests which were using the old one.