wifi: Allow connecting to WPA2 networks
As it turns out, there is nothing different to do when actually
connecting to the network, the only problem was properly checking
whether the network was "secure".

We were ignoring the RsnFlags attribute, which is the one containing
security information for WPA2 networks (not WpaFlags as for WPA1
networks).

And as a result, we just wouldn't display the popup asking the Wi-Fi
key, which means the site admins would never manage to connect.