debian: Ignore the Makefile
The Debian packaging machinery was finding the Makefile, and as such it
decided to run "make" in order to package ideascube.

This is wrong, it should use the setup.py to build the package.

This commit tells the Debian packaging machinery to ignore the Makefile,
and instead use the setup.py-based Python packaging.