wifi: Don't pretend we connected when we didn't
Because NetworkManager is asynchronous, when we try to connect all we
actually do is ask him to connect. We fire the request and then assume
it's done.

But if something went wrong (e.g we supplied a wrong passphrase), then
NM will not actually connect us.

We'd never catch that error, and just tell the user everything had gone
fine.

This commit makes it so that we wait a bit after firing the conneciton
request, and only report success if we find that we actually got
connected.

This is clearly not how we should be doing this, we should use the
asynchronous interface as it was intended, listening on signals.

But this has the merit to work right now, and to improve the current
situation where users got very confused when they typed a wrong
passphrase.

Fixes #254