Move the extra files out of the debian/ directory
Those files are useful when building a Debian package, since they get
added to the package to make the out-of-the-box experience of installing
the ideascube package just work.

However, they are also useful in other contexts, for example a Fedora
package, or even people installing it manually.

Therefore they shouldn't be hidden in the debian/ directory.

Since they are additional, optional things to the actual sources though,
they will be better in a new extras/ directory.