extras: Add the systemd service files
For the Ideascube serveradmin, we need the application to run as a PAM
session (see previous commit for the PAM configuration) so that it can
control the network through NetworkManager.

We can do that simply with the appropriate systemd configuration
directives:

    User=
    Group=
    PAMName=

However, before we can use these, we need to run uWSGI as a native
systemd service.

Unfortunately, on Debian Jessie uWSGI is still run as a SysV init
script.

Therefore, we add here a uwsgi.service file to run uWSGI as a native
systemd service in addition to the uwsgi.service.d/ideascube.conf file
we really need.

If uWSGI ever gets migrated to a native systemd service in Debian, we
will be able to just remove this file.

Fixes #267