debian: Move the postinst script to systemctl
Let's use "systemctl restart foo" instead of "service foo restart",
since that's the preferred way on Debian going forward. If they ever
remove the "service" command, then we won't be surprised.

In addition, we need to use tell systemd to reload the files on the
system, because our package installs some .service files.