catalog: Represent remote sources
We are implementing a content manager, which will be able to download
and install apps and content coming from remote sources.

This is the first step towards that goal, representing those remote
sources as an object, storing it on the file system to remember it, and
loading a previously configured one from the file system.