catalog: Implement remote source management
We are implementing a content manager, which will be able to download
and install apps and content coming from remote sources.

This commit introduces the Catalog object, which is going to be our
actual package manager.

For now, all it can do is list known remote sources, add new ones and
remove known ones.