catalog: Implement cache management
We are implementing a content manager, which will be able to download
and install apps and content coming from remote sources.

This commit improves the Catalog object, so that it is capable of
downloading from each of its remote sources the metadata representing
the content they make available, and keep it as local cache.

We can also clear the local cache, at which point no content seems to be
available any more, until the next cache update.

This commit also brings in a (completely untested) method to print a
pretty progress bar for downloads. Eventually this needs to be
implemented better, maybe plugged to a more central reporting in
ideascube.