catalog: Represent package handlers
Just like we will have various types of packages, each with its way of
being installed, we will also have various types of package handlers.

For example, we could imagine installing zipped zim packages for Kiwix,
or for another application consuming the same file types.

This commit introduces the concept of package handlers, with a base
Handler class and a dedicated Kiwix handler (since that's what we're
about to implement)

The Handler base class has a registry of known handlers, much like we
already do for packages.

Handlers can install and remove packages, by simply asking the package
to install or remove itself, passing it the right parameters.