catalog: Allow specifying additional package caches
Our users often have pretty terrible networks. As such, downloading a
40GB Wikipedia archive might be very hard for them.

This gives them an additional possibility: if they somehow managed to
get the package (downloading it from a place with a faster connection,
taking it from the cache of an other Ideascube deployment, etc...), they
can now specify the folder containing this package as an additional
package cache.

If they do that, then we won't download the package at all when trying
to install/update it, we'll just take it from this additional package
cache.

Because sometimes, in some places, physically sending a USB disk leads
to a higher bandwidth than actual downloads.