catalog: Implement package installation
Finally, our catalog can now install packages available in a remote
source.

This requires a few bits to be in place, though.

First, we add the ability to download a package, and even resume a
previous incomplete download, which is great for our users and their
slow/unstable connections.

Downloaded packages are stored in a local package cache on disk, where
they can be reused if needed, to avoid the need to download again. Disk
space is much cheeper than network bandwidth for our users.

The local package cache part looks a bit weird right now, as if we could
have more than one? That's going to make sense suddenly in a future
commit.

The catalog receives a list of package ids, and from that it
instantiates the corresponding Package objects, the proper Handler
object(s) for these packages, then ask the handler(s) to install the
packages, and finally to run their post transaction tasks.