catalog: Implement package listing
And now our catalog can list installed, available or upgradable
packages.

The catalog receives a list of package ids, and from that it
instantiates the corresponding Package objects and returns them.

That's not super useful though, because you need to provide the ids of
the packages to list explicitly.

If only we supported globs...