catalog: Implement package removal
And now our catalog can remove packages available that it had installed
previously.

The catalog receives a list of package ids, and from that it
instantiates the corresponding Package objects, the proper Handler
object(s) for these packages, then ask the handler(s) to remove the
packages, and finally to run their post transaction tasks.