catalog: Support globs for all package operations
Having to pass the ids of the packages we wanted to list was not very
useful, because you need to list packages to know their ids...

This commit introduces the possibility to use wildcard globs when
listing packages, so that we can, say, list all available packages with
ids like 'wikipedia.*'.

This also finally allows us to list **all** packages, simply by specify
the list of ids to list as ['*'].

And because of the way the code is organized, without doing anything
more, we get support to install, remove and upgrade packages with globs.

So we can now install all the packages for a given locale:

    catalog.install_packages(['*.tum'])

Or we can even upgrade all the packages at once:

    catalog.upgrade_packages(['*'])