Handle MediaCenter packages
With this, we can install media content with the package manager, and it
gets imported into the Ideascube media center.

Since we do not want to backup these media files (they can be
reinstalled with the package manager at any time, after all), we trick
Django into thinking the files are in its MEDIA_ROOT with a symlink
(symlinks aren't backed up).

We also need to use a custom form to validate the datas without using a
FileField that would copy the file.