Drop support of creating zip backup.
Zip format is generally a pain to deal with (hello non-ascii
characters).

In addition, we're about to add filtering of which files to backup and
which files to ignore, which is trivial to do with the tarfile module,
but much harder with the zipfile one.

To keep things simpler, we've decided to remove Zip support.

Zip backups are still supported for restore, though.

[CONFIG_CHANGE]
- zip is no more supported for BACKUP_FORMAT.
- Default value for BACKUP_FORMAT as change from 'zip' to 'gztar'