Properly set available_kinds and available_langs in views.
The available kinds must not be kinds for whom there are documents.
This must be kinds for whom there are documents *using other filters*.
(Same apply for lang)

We also move all the get_context_data and get_queryset in a
FilterableViewMixin and lets views on filterable content us it.
(blog, library and mediacenter)