Do not fail if we try to add the same remote twice. (fix #357)
If we try to add a remote already existing (same id) with same url,
print a warning.
If we try to add a remote already existing but with different url,
raise as usual.

The Catalog.add_remote always fails. This is the add command who checks
check the exception and print only a warning if necessary.