Split the catalog cache
Today the catalog caches a few things:

  1. the list of configured remotes
  2. the list of packages available in each remote
  3. the downloaded package files
  4. the list of installed packages

Items 2 and 3 are clearly cache: if they get dropped, then they can be
obtained again by updating the cache of the catalog or redownloading the
package(s).

However, items 1 and 4 are really local data, and as such they would
need to be backed up. This way, we could only backup the fact that a
package is installed and not its actual contents, and when restoring
from a backup we'd just have to reinstall the package.

So we need to split the current catalog cache, and move some of it to
a more permanent storage.

Relates to #376