Work around test brokenness
Both our databases are supposed to be in memory when running the tests.

Unfortunately, the transient ones is on the file system, under
STORAGE_ROOT in two instances at the beginning of the tests:

* when running the migrations
* when firing the post_migrate signal in ideascube.search.apps

This wouldn't have been a problem back when we were using the dev
config as a fallback, since that config created STORAGE_ROOT.

However, we now fallback to the base config, which doesn't try creating
its STORAGE_ROOT.

As a result, the tests are failing, because the transient db couldn't
be created, because its parent folder didn't exist.

Explicitly setting the TEST_NAME in the config file to ':memory:' doesn't
change anything.

But somehow, doing it in pytest_configure does the trick.