Introduce a helper to get the server name
Lots of places in the code were directly reading the IDEASCUBE_NAME
configuration entry.

We're eventually going to move away from that setting, and have it
configurable in the server admin interface.

This commit prepares for that, as now all the code requiring the server
name go through that single helper function.