Use an in-db Setting for the server name
Having this in the configuration files is not ideal, among other reasons
because it forces the ansiblecube folks to create a setting file for
every single deployment, even if just for this.

Storing it as a Setting in the database allows us to drop this setting
from the configuration files entirely.

However, the already deployed boxes have their names defined in their
respective configuration file. We don't want them to upgrade and
suddenly visitors see the default name displayed instead of the one
their admins had chosen.

So when the site name hasn't been configured locally in the database,
the code falls back first on a name from the configuration file, and if
there isn't one then on a default, hard-coded name.