Add an in-db configuration management system
We're going to move stuff from the settings file to forms in the admin
UI, so we'll need to store configuration into the database.

This introduces a new model that just that.

Each entry in the table is a configuration option. It has:

* A `namespace`, to group together settings related to a particular part
  of ideascube. For example, settings related to the media center would
  all have the `namespace` set to the 'mediacenter' string.

* A `key`, which is the setting name. For example, the configuration
  option to control the site name could have its `key` set to the
  'site-name' string.

* A `value`, which is the configuration value. It could be anything, as
  long as it is JSON-serializable, as that's how it gets stored in the
  database.

* An `actor`, which the person from the staff who made the configuration
  change.

* A `date`, which is the datetime at which the configuration change
  occured.

I'm adding a convenience API with `set`, which allows to easily set the
value of a setting, whether it already exists or not, abstracting the
underlying model API.

I'm also adding another convenience API with `get_string`, which allows
to get the value of a setting, eventually providing a default value.

I'm only adding this for string settings, because that's all I'm going
to use in the short term, adding similar ones for other types is left as
an exercise for those who will need them in the future.

Implements #404