Add a django command to create a package from a csv (#397)
* Add a django command to create a package from a csv.

The csv has the same format as the csv used by import_media.
The created package should be added to a repository to be installed later.

* move csv2pkg command to serveradmin app

* Handle case where preview in the media manifest is defined but empty.

* Add the preview file in the package if preview is defined.

* Be sure that tags are properly indexed. (#405)

* Add a dry-run option to avoid really create the package.

This allow to test the validity of the metadata without creating the pkg.

* Fix few remarks on made during reviews.

Also removed unused calculation of size and sha256sum.
(This code where initially written in the context of creating the catalog
in a all. If we create only the package, we don't need these attributes)

* Rename command csv2package to csv2pkg.

* update help of csv2pkg command