build: Don't use wheels
We want, and need, to rebuild everything from source, without using
pre-built wheels.

One reason is that it's just the right thing to do, but it's especially
pressing because the build actually fails with wheel, as described here:

    https://github.com/spotify/dh-virtualenv/issues/150

The solution would be to pass the --no-binary=:all: option to pip,
except it appeared in pip 7.0.0, and we're stuck on a much older one.

Fortunately, theold version had --no-use-wheel which does the same.

This commit bumps the dh-virtualenv dependency to 0.10 (only in the
Jessie backports), so we can use its --pip-extra-args option to tell pip
not to use wheels.

Relates to #408