Allow filtering which languages appear in selectors
When creating new content (blog article, media, books), one can specify
which language the content is in.

We don't really want to show a big list of all the languages in the
world.

And in fact, we used to only show a list of configured languages.

However, with the previous commit, we're now displaying them all.

The problem is that we want the model to **accept** all languages, but
we want to **display** only some of them.

We can't base ourselves on the AVAILABLE_LANGUAGES setting as we used
to, because that actually represents the languages Ideascube itself is
available in.

This commit introduces a new in-database configuration which stores the
list of languages to allow in the various forms, and falls back on a
hardcoded list of a few common languages.

Relates to #425