Test all the config files
We regularly make little syntax errors in the config files, leading to
lots of pain down the road.

With this test, config files will be validated at each push to the
ideascube repository, which should make us happier.

And in fact, it actually found a broken config!

That config is old, and probably worked at the times of Python 2, but we
moved to Python 3 since then. Good thing that box was never updated. :-/

Fixes #395