Remove reindex migration
This migration was problematic.

First of all, it was only ordered after the initial search migration
(which creates the model). But that one wasn't ordered after anything at
all.

We've been lucky so far that Django somehow always was running those 2
migrations after everything else, so things just worked.

But it's very possible that they could have run before other migrations,
and in fact it's very easy to do: just add a new migration (for an app
different from search) and make it depend on 0002_reindex: the two
search migrations suddenly get bumped by django much earlier in the
migration cycle.

And this is a problem.

Assume a new migration adds a column to the Book model.

If that migration runs after the 0002_reindex migration, then the
0002_reindex migration will try indexing the new column in the Book
model, which doesn't exist yet as the book migration hasn't run.

As a result, we'd need to actually add dependencies to the 0002_reindex
migration, and we'd need to constantly bump them to always run after all
other migrations, even ones we'd create in the future.

That's just too much work.

But we already have a post_migrate signal handler which reindexes the
database! That means we always reindex anyway after all migrations have
run.

As a result, we just don't need 0002_reindex any more, so we should just
get rid of it.

However, there is one thing it was doing which is useful: deleting the
old table from the default database. This is preserved in a new
migration.