Add a card template to display media_center packages.
Some package instal media, and we need a card for this package.
As a card need a link, we need a link to display only the media comming from this package.

So, we need to filter media on the package_id.
We add a source attribute on the search model to be able to filter on it.
The source attribute is only use by media with the package_id as value for now.