Update Mullah Piaz card
We can not use the kiwix pre generated card as the path to the Mullah
Piaz file is not the same.
The file is mullah-piaz-digest_fa_all_2012-12.zim and result in being
mullah-piaz-digest.fa
In reality, the website answer to
http://kiwix.ideasbox.lan/mullah_piaz_digest/

We should understand why this file change during unziping, but for now I
fix the card.