configuration: Add a registry of all known configs
This will be useful to tighten a bit the API and avoid mistakes.

It will also be very useful when we want to implement a command-line
tool to manage the configuration.

For now, the registry only contains the namespaces and keys we already
use as settings, more will come eventually.

Relates to #441