configuration: Allow getting configurations
This is the big missing part of the API: we already could store
configurations in the database, now we can retrieve them.

This builds up on what was done previously to ensure the correctness of
the configuration in the database, based on the registry. In addition,
it ensures that it will never return a value of an incorrect type, even
if the database somehow got corrupted.

In this way, consumers of the API can always trust the configuration
system to preserve the integrity of the database, and they can do away
with verifying things.

Relates to #441