Completely hide the system user
This is a hidden user, to be used exclusively by Ideascube itself.

There is no point showing to the admins, since what they are interested
in is the list of **people** having an account on the box.

However, there are cases where we will need to be able to get that user,
so this commit leaves an API for this, to be called explicitly.