Improve getting the system user
Having a way to get **really** all users, including the system one is
useful, but what we'll want most of the time is to get the actual system
user.

This commit introduces a manager method to do just that.

Credit for the idea goes to @mgautierfr, who made me see the light here.