catalog: Fix some broken assumptions
We were assuming too much about the structure of zipped zim packages.

Specifically, for a package downloaded at:

 http://download.kiwix.org/portable/other/kiwix-0.9+bouquineux_fr_all_2013-07.zip

We assumed that the zim files inside it would always be named:

 bouquineux_fr_all_2013-07.zim*

But that's not the case, they were named:

 bouquineux_07_2013.zim*

This commit drops that assumption, and instead tries harder to find the
zim files as they are named.