Correctly catch output from logging to stderr.
The default Handler from logging store "in cache" the sys.stderr when it's
create.
When capsys from py.test try to mock sys.stderr, it is to late.

But logging has a Handler who always print to sys.stderr and do not cache
it.

We configure logging to use it, then, mocking from py.test works correctly.