Add favicons
This commit adds favicons for Ideasbox and Koombook, made with love and
my incredible design skillz. In other words, they should be replaced as
soon as possible by real artwork made by someone with a clue.

Ideascube choses the favicon to use based on the IDEASCUBE_BOX_TYPE
setting.

Favicons are under the static root, so they can be served by nginx and
cached by the clients.

Relates to #466