Drop the javascript view
This was a view, and the related URL, which had a template which it
rendered as a Javascript response containing... almost nothing.

And that view was called on every single page.

Sure, we can cache it.

But we can also remove it completely, inline it into the base template,
and clients just won't do ther equest at all.

Better than a cached request: no request.

Relates to #469