mediacenter: Redo guessing the document kind
We were trying to guess the document kind when saving the document, but
only if it had an 'original' file and if the kind was set to 'other',
based on the 'original' file name.

And then, right after, we would always overwrite that by trying to guess
the kind from the uploaded 'original' file content type.

This commit reworks that, simplifying it greatly and making it do what
we actually want: try guessing if the kind was set to 'other' in the
form, based on the content type if possible and on the file name if not.

In addition, this really is part of validating the user-supplied 'kind'
field from the form, so all this code is moved to the appropriate clean
method.