Allow aggressively caching media files
This commit introduces a 'media' template tag, which automatically
returns the URL of a media file from a model instance, and suffixes it
with a query string containing the last modifcation time of that model
instance.

The new tag is then used everywhere in the appropriate templates.

This allows us to configure nginx to tell the client to cache the media
files for a much longer period of time, so that much less requests
actually get sent to the server.

Fixes #468