Remove a useless test
This test function tries to ensure that we can upload a document without
specifying the content-type of the file being uploaded.

Except it doesn't do that at all, because WebTest's Upload helper will
automatically set one.

Therefore, this test function is completely redundant with the other
ones, in addition to being confusing.