Display cards for installed packages.
The card's template is mainly determined from the package type (kiwix, media,
static-site). (Except for static-site this is a bit more complex and we use
the template_id to find the template to use.)

As consequence, a lot of templates are removed and replace by the generic one
kiwix.html. Specific information for this template (theme and css_class) are
determined from the package id.
Those information should be found in package's metadata/settings/... but for
now, we don't know hom. So they are in the code.

Cards corresponding to packages have been remove from config files as they
will be displayed when packages are installed.