search: Recreate the index table before migrating
We have two problems with the search app currently.

The first is that if an instance of ideascube gets updated from a very
old version, the migrations fail. (#456)

The second is that because the Search model is not managed by Django, if
we add new fields to it then no migration gets created, which breaks the
application as well as the post-migrate reindex. (#489)

The solution to both is actually quite simple: simply always drop and
recreate the index table before running the migrations.

Fixes #456
Fixes #489