Do not follow unnecessary redirection in fixtures.
loggedapp and staffapp fixtures submit the form to log the user.
The redirection to the index page is useless in our case.

Moreover, it imply the generation of the index.html page who may run a
lot of code.

As it not needed and not wanted, remove the follow() call.