Don't rely on an unguaranteed order
Django offers no guarantee of the order in which the fields of a form
are cleaned.

As a result, we can't have a clean_foo method relying on the fact that
another field bar has already been cleaned.

For this kind of cases, we must use the clean method, which is what
we now do with this commit.