library: Redo the model
People found it too annoying to have to create a book (and its
specimens) both in the stock and in the library.

We listened to them, and we're making the library models just inherit
from the stock management ones.

This means that creating a book/specimen in the library will
automatically make it available in the stock management app.

This commit adds a migration to try and ensure that all existing books
and specimen are turned into the new model.

Left for a future iteration is how to deal with books created directly
in the stock management app.