serveradmin: Use Cockpit to set the timezone
This tries to use Cockpit [1] to talk to the timedate1 daemon through
DBus through a WebSocket.

The list of timezones is obtained the same way systemd does it, so that
should work properly.

The view gracefully degrades to a « not available » message if going
through Cockpit doesn't for for any reason.

[1] http://cockpit-project.org/