catalog: Update the installed metadata
We've recently started using the installed metadata of the catalog to
display things that are visible to the user, mainly the cards on the
home page.

However, that made us realize that our user-facing metadata was pretty
bad. So we've fixed it on the remote.

But we need a way for clients to update this metadata, without having to
reinstall the packages.

With this commit, every time we update the cache of available packages
we will also update the metadata of the installed packages, as long as
none of the sha256sum, type and version keys changed.

Fixes #513