Simplify the installation of static site packages
We were creating an NGinx configuration file for each package. Among
other things, it meant the packages needed to be installed as root.

This changes all that, so that we now have a single vhost for all static
sites, pointing to the installation root of static site packages, and
serving them all.

We can thus drop lots of code from the Nginx handler's commit function,
making package installation faster and more robust.

Fixes #431