[WIP] Add a html/js proxy page to add button on top of a kiwix page.
This is need proper configuration of nginx to redirect any
kiwix.{{domain}}/book/.* to the the static page.

kiwix-serve need to be launch with `--nosearchbar` to prevent it to add
the search bar itself.

This is a WIP/RFC this is not finish (no autocomplete of the search,
styling is missing).