Redo the branding management
This introduces a new top-level directory in the static files which
contains all the resourcesto brand a box.

By default, it contains artwork for the Ideascube project itself.

This means we don't load the Ideasbox or Koombook artwork any more. I'm
going to bring it in a future commit, but differently.

This also means that we can completely get rid of the IDEASCUBE_BOX_TYPE
setting, since it was only used for the branding.

Relates to #522
Fixes #302
Fixes #466