Add an automagic TextIOWrapper
It tries to automatically detect the encoding and just wraps a
file-like (BytesIO) into a stream of decoded strings, as they are
iterated on.

For now, it will assume that everything which isn't UTF-8 is Latin-1,
since those are the two common encodings we've met in the field.

If that ever becomes a problem, we can move to something like cchardet.