Remove test of the JsonField.
Having a model specially for the tests without migration script cause
tests to fail if run with migration. As there is no migration script,
the table is not created and test fails.

We agreed that the test of the JsonField do not test a lot as :
 - The field itself just call json.dumps, json.loads
 - The field is already (integration) tested with test of configuration.

So we've decided to remove it.